osd: kludge to efficiently rebuild past_intervals in parallel on startup
Particularly tortured clusters might be buried under thousands of osdmap
epochs of thrashing with thousands of pgs.  Rebuilding the past_intervals
becomes O(n^2) in that case, and can take days and days.  Instead, do the
rebuild for all PGs in parallel during a single pass over the osdmap
history.

This is an ugly (mostly) one-time use hack that can removed soon.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>