objclass: provide new api for unique subop versioning
We need to be able to generate a unique identifier for
each subop. This can be useful e.g., if we want to keep multiple
omap data entries indexed by the same key. The unique id
is being generated by the current object id and the current
osd subop count. For write operations it's unique.

Signed-off-by: Yehuda Sadeh <yehuda@inktank.com>