inttypes: detect and define missing integer types
Working around missing integer types is pretty easy. For example, the
__u32 family are Linux-specific types, and using these in Ceph
internally is fine because we can typedef them, but the biggest problem
is that they are also used in installed headers. Again, they could be
typedefed in the installed headers, but it isn't clear that we can't
just use standard ints in installed headers.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>

Conflicts:
	src/include/byteorder.h