FileJournal: zero-fill in-lieu of posix_fallocate
Zero-fill journal if posix_fallocate fails or if it is not supported.
For very large journals zero fill can take a long time. An optimization
is to write a zero byte to the end of each block, or use platform
specific features for file allocation. Reference solutions for various
platforms can be found in Mozilla, SQLite, and PostgreSQL.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>