ObjectCacher: store image_overlap in ObjectSet
Pass the image_overlap to the WritebackHandler functions that need it,
may_copy_on_write() and write(), so that they don't need to take any
locks to read the current overlap. This is a step towards changing
the lock ordering in librbd so the cache lock may be taken inside
of the snap and parent locks.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>