os/FileJournal: trigger another commit if still > 1/2 full after a sync
Previously we tried to be clever and only signal when we just crush the
half full threshold.  This would fail to trigger a second commit if the
sync didn't drop us back below 1/2 full.

Fix this by signaling a second commit if we are still more than 1/2 full
after the previous sync completes.

Signed-off-by: Sage Weil <sage@redhat.com>