No version on ceph-libs Obsoletes.
If we are installing with the new package structure we don't ever want the
new package to co-exist with the old one as the install would fail anyway.
Since there are already I don't really see any benefit to having a version
listed where it will cause breakage if the version isn't the newest one.

There is currently a 0.81 package and thus < 0.80.5 caused collisions.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>