msg/SimpleMessenger: do not block for fast dispatch in mark_down[_all]()
Currently the mark_down methods call stop_and_wait(), which means we block
while any in-progress fast dispatch completes.  This will deadlock if we
are holding any locks that a fast dispatch handler may need.  (For one such
deadlock, see #9462.)

The original motivation for blocking on mark_down was to avoid clearing
priv on cons while message handlers were in flight, but we recently made
this safe to do (as there are other paths that can clear it out at any
time, notably require_same_peer_instance() which is called from both
fast and regular dispatch).

So, stop blocking on mark_down[_all]().  Keep blocking for session takeover
in the connection race handling as it may still be necessary there
(although I suspect there are still some issues there).  This kicks the
can a bit further down the road, but moves us in the right direction, I
believe.

Fixes: #9462
Signed-off-by: Sage Weil <sage@redhat.com>