mds: CInode: CDir: add scrub_[stamp|version] handling
TODO: Is resetting the scrub stamps on pop_projected_link() sufficient for
clearing it out on every rename?

Give each CInode a stamp and version for both the last started scrub, and
the last completed scrub. The last started scrubs are ephemeral, but we
add infrastructure to track when the "complete" values aren't flushed out
to disk.

Signed-off-by: Greg Farnum <greg@inktank.com>