mds: CDir: add more scrub metadata, and setup/handling functions
We don't want to force inodes-to-scrub to stay in the cache, so don't
use pointers to CDentry, just use dentry_key_t. And provide them for
non-dirs as well as dirs. And provide a list of those that we have
scrubbed in addition to those we haven't.

Implement setup_scrubbing() to initialize these structures, and a
mark_and_get_next_scrub_dentry() to move dentries between these lists and
get the next dentry to scrub.

Signed-off-by: Greg Farnum <greg@inktank.com>