osd/ReplicatedPG: writeback: always forward reads, even if we promote
If we are about to promote based on a read op, forward the read first. That
op will get a lower latency response, and it won't slow down subsequent
ops significantly since it'll just prime (some of) the cache for the
promote.  If we are lucky, there isn't a second read right behind us that
will block on the promotion.

Signed-off-by: Sage Weil <sage@redhat.com>