rgw: use /var/run/radosgw for FCGI socket
The radosgw init scripts run radosgw as the "apache" (RHEL) or
"www-data" (Debian) unprivileged UID. The default socket directory
(/var/run/ceph) is only writable by root, so the unprivileged radosgw
daemon was not able to write a socket file to /var/run/ceph.

Add a new /var/run/radosgw directory to the packaging and make it owned
by apache (RHEL) or www-data (Debian).

Update the documentation and sample RGW configs to use the new
/var/run/radosgw. Remove the section about chown'ing /var/run/ceph to
Apache, since this is no longer required.

Remove the step in the radosgw upstart configuration that would mkdir
/var/run/ceph, since radosgw no longer uses that location.

http://tracker.ceph.com/issues/9001 Refs: #9001

Reported-by: Alexandre Marangone <amarango@redhat.com>
Signed-off-by: Ken Dreyer <kdreyer@redhat.com>