common, global: use lttng ust functions for handling fork behavior
This allows lttng to work across fork(2) and related calls, including
daemon(3). Without these calls, LD_PRELOAD would need to include
liblttng-ust-fork.so, which wraps fork and daemon.

Without these wrappers, the atexit handlers registered by lttng may
cause a crash in when a forked process exits. This does not happen
with upstart, since it runs ceph-osd and other daemon in the
foreground.

Signed-off-by: Josh Durgin <jdurgin@redhat.com>