Add Accelio Submodule
This change just adds the submodule itself.  Note that Accelio's
master branch is regarded as stable.

"""
This is a build/packaging issue.  I think right now libxio needs to be
present at build time and it's not available in any of the distros we
build for.  Usually what we do in this situation is embed it as a git
submodule (like we do with rocksdb).  Happy to take patches that do
that\ if that's the best route (I suspect it is?).
<snip>
sage
"""

Signed-off-by: Matt Benjamin <matt@cohortfs.com>