PROTOTYPE mon: maintain a list of ProgressEvents and output in status
Not doing any persistence of anything: the state for these guys just
lasts as long as the mon leader is running, after the time the
event is instantiated.

Just tag the output onto "ceph status" for simplicity, a finished
version of this would need its own output function, and to generate
some cut down version for use in "status", perhaps by simply outputting
the number of activities in progress and the aggregate %ge completion?

Signed-off-by: John Spray <john.spray@redhat.com>