Flusher respects sync gen numbers, so will not flush entries while entries from a prior sync gen number (before the latest aio_flush()) are still flushing.
When dispatching writes in persist-on-write mode, a sync point will be inserted
before the write if any writes have completed in the current sync point. So, once
a write completes in persist-on-write mode subsequent writes will not be flushed
concurrently with it because they will have a higher sync gen number.