Remove block alignment checks on incoming requests.
This is an image extent based interface, and while we have a minimum
allocation size for RWL entries, incoming requests can be any size or
alignment.

Translation functions between block and image extents are retained,
since these are encoded differently, but the notion of a block size
has been removed.