RWL::invalidate() now flushes first by default.
If you really want invalidate to discard unflushed writes, there's
a version of invalidate() that accepts a "discard" flag.

Invalidate seems to only be called by the snapshot code. The uses of
invalidate there seem to assume the image cache is just a read cache,
and don't specifically need unflushed writes to be discarded.