Preparing for aio_writesame()
Write path generalized to reuse most of it for writesame. The alloc_write_resources()
function has been moved to a class method of C_WriteRequest.

Added specific WS requests, log op, and log entry. Resource allocation for
WS happens, then we dispatch the request as if it's a normal write. That
won't work until the write dispatch path is generalized the same way.