Write and WS prepare pmem_bl on entry create, then use substr_of() in read
These log entries now contain a buffer::ptr that's initialized with
buffer::create_static() when the entry is cerated. We no longer
create buffer::raw's from that at read time, and no longer create
deleters every time we do that. We'll use buffer::ptr::raw_nref()
to detect when all the readers have released their references on the
pmem buffer.

This still needs to be applied to the flush path, and laod_existing().